YABD-3 Do refactoring
