+++ yabd/plugin/src/main/kotlin/dev/yabd/plugin/internal/core/utils/JiraUtils.kt
@@ -1,10 +1,10 @@
 package dev.yabd.plugin.internal.core.utils
 
-import dev.yabd.plugin.internal.usecase.jira.model.request.BodyNetModel
-import dev.yabd.plugin.internal.usecase.jira.model.request.ContentCommentNetModel
-import dev.yabd.plugin.internal.usecase.jira.model.request.ContentNetModel
-import dev.yabd.plugin.internal.usecase.jira.model.request.JiraCommentRequestNetModel
-import dev.yabd.plugin.internal.usecase.jira.model.response.JiraFileUploadResponseNetModel
+import dev.yabd.plugin.internal.data.jira.model.request.BodyNetModel
+import dev.yabd.plugin.internal.data.jira.model.request.ContentCommentNetModel
+import dev.yabd.plugin.internal.data.jira.model.request.ContentNetModel
+import dev.yabd.plugin.internal.data.jira.model.request.JiraCommentRequestNetModel
+import dev.yabd.plugin.internal.data.jira.model.response.JiraFileUploadResponseNetModel
 
 object JiraUtils {
     fun getDownloadLinkComment(comment: String): JiraCommentRequestNetModel {
