+++ yabd/plugin/src/main/kotlin/dev/yabd/plugin/internal/data/jira/model/response/BodyNetModel.kt
@@ -1,4 +1,4 @@
-package dev.yabd.plugin.internal.usecase.jira.model.request
+package dev.yabd.plugin.internal.data.jira.model.response
 
 import com.google.gson.annotations.SerializedName
 import dev.yabd.plugin.common.model.base.NetModel
