+++ yabd/plugin/src/main/kotlin/dev/yabd/plugin/internal/data/slack/model/response/SlackFileResponseNetModel.kt
@@ -1,4 +1,4 @@
-package dev.yabd.plugin.internal.usecase.slack.model
+package dev.yabd.plugin.internal.data.slack.model.response
 
 import com.google.gson.annotations.SerializedName
 import dev.yabd.plugin.common.model.base.NetModel
